Add cart test

+++ shared/models/cart.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/lucsky/cuid"
 	"github.com/jinzhu/gorm"
+	"github.com/synergydesigns/stylesblitz-server/shared/utils"
 )
 
 type Cart struct {
@@ -59,7 +60,7 @@ func (service *CartDBService) CreateCart(userID string, vendorID string, cartTyp
 		foundProduct := Product{}
 		result := service.DB.Where("id = ?", cart.TypeID).Limit(1).First(&foundProduct)
 		if result.Error != nil {
-			log.Printf("An error occurred creating cart %v", result.Error.Error())
+			log.Printf("An error occurred creating cart.\n %v", result.Error.Error())
 
 			return nil, fmt.Errorf("an error occurred creating cart %s", result.Error.Error())
 		}
@@ -74,8 +75,14 @@ func (service *CartDBService) CreateCart(userID string, vendorID string, cartTyp
 
 	if cartResult.Error != nil {
 		log.Printf("An error occurred creating cart %v", cartResult.Error.Error())
+		fmt.Println(cartResult.Error.Error(), "from create cart")
 
-		return nil, fmt.Errorf("an error occurred creating cart %s", cartResult.Error.Error())
+		if utils.ForeignKeyNotExist(cartResult.Error) {
+			return nil, fmt.Errorf("Vendor with id %s does not exist", vendorID)
+		}
+
+		fmt.Println(cart)
+		return nil, fmt.Errorf("see here an error occurred creating cart %s", cartResult.Error.Error())
 	}
 
 	return &cart, nil
@@ -116,11 +123,14 @@ func (cart *Cart) CartType() (CartType, error) {
 }
 
 func (service *CartDBService) DeleteCart(userID string, cartID string)(bool, error) {
-	cartResult := service.DB.Delete(&Cart{}, "id = ?", cartID)
+	cartResult := service.DB.Delete(&Cart{}, "id = ? AND user_id = ?", cartID, userID)
 	if cartResult.Error != nil {
 
 		log.Printf("An error occurred deleting cart %v", cartResult.Error.Error())
-		return false, fmt.Errorf("An error occurred deleting cart %s", cartResult.Error.Error())
+		return false, fmt.Errorf("An error occurred deleting cart. %s", cartResult.Error.Error())
+	}
+	if (cartResult.RowsAffected < 1) {
+		return false, fmt.Errorf("An error occurred deleting cart")
 	}
 
 	return true, nil
